sun now doesnt not render 3 times
