+++ server/public/index.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
 
   <title>Solar System</title>
-  <link rel="stylesheet" href="/styles.css" />
+  <!-- <link rel="stylesheet" href="/styles.css" /> -->
 </head>
 <body>
   <div id="app"></div>
